requirements.txt
azure-rest-ai-inference
azure-core-auth
azure-core-sse
requests 